requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.4,<3.10.0
django-mysql==2.2.0

mysqlclient==1.3.12

flake8>=3.6.0,<3.7.0